Update dfrobot-weatherstation-rpi.py
Rewrote original code, as there was possibility to only get a partial read of the 37-byte string in the original method. This method waits for the first occurence of the "c" in the 37byte string reads through the data 4 bytes at a time (where appropriate the last 2 elements returns get read at 2 and 6 bytes) and returns the data in that way. The lists that are returned are converted in all of the for loops to ints. Next up, I'll either ditch the Apache HTTP element and just do it all in Python or maybe rewrite this to be a realtime socket service.